pyproject.toml
[tool.poetry]
name = "mediwave-policies-lang-rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
rag-weaviate = {path = "packages/rag-weaviate", develop = true}
redis = "^5.0.1"
langchain-community = "^0.0.21"
gpt4all = "^2.2.1.post1"
python-dotenv = "^1.0.1"
retry = "^0.9.2"



[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
